requirements.txt
openai
gradio
gitpython
pillow
requests
python-dotenv
PyMuPDF
pyyaml
